scrapy_crawler/requirements.txt
scrapy
psycopg2-binary
python-dotenv

